Refactoring the code and merging the functionality of Map.and TiEs. into a single component
