--- client/src/components/Dashboard.css
@@ -1,60 +1,0 @@
-
-.Dashboard {
-  display: flex;
-  flex-direction: column;
-  justify-content: center; 
-  align-items: center; 
-  min-height: 100vh; 
-  padding: 5rem;
-  background-color: #f4f4f4; 
-}
-
-.Dashboard-box {
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-  width: 80%; 
-  max-width: 1200px; 
-  padding: 2rem;
-  background-color: white; 
-  border-radius: 10px; 
-  border: 2px solid #ddd; 
-  box-shadow: 0 4px 10px rgba(0, 0, 0, 0.1); 
-}
-
-.Dashboard h2 {
-  font-size: 2rem;
-  margin-bottom: 1rem;
-  color: #333;
-  text-align: center;
-}
-
-.Dashboard .loading-message,
-.Dashboard .error-message {
-  font-size: 1.2rem;
-  color: #333;
-  text-align: center;
-}
-
-@media (max-width: 768px) {
-  .Dashboard-box {
-    width: 95%;
-    padding: 1.5rem;
-  }
-
-  .Dashboard h2 {
-    font-size: 1.8rem;
-  }
-}
-
-@media (max-width: 480px) {
-  .Dashboard-box {
-    width: 100%;
-    padding: 1rem;
-  }
-
-  .Dashboard h2 {
-    font-size: 1.5rem;
-  }
-}
