+++ client/src/components/NavBar.css
@@ -17,36 +17,42 @@
   font-size: 2rem;
   font-weight: bold;
   text-decoration: none;
-  color: #007bff; 
+  color: white; 
   letter-spacing: 2px;
-  background-color: rgba(255, 255, 255, 0.6); 
+  background-color: black; 
   border-radius: 10px;
   padding: 10px 15px;
   text-align: left; 
   width: 200px; 
+  
 }
 
 .navbar-links-left {
   display: flex;
   justify-content: flex-end; 
   gap: 15px; 
-  padding-right: 20px;
+  box-sizing: border-box;
 }
 
 .nav-link {
-  font-size: 1rem;
+  font-size: 1.2rem;
   text-decoration: none;
-  color: #007bff; 
+  color: white; 
   transition: color 0.3s ease;
   display: flex;
+  justify-content: flex-end;
   padding: 10px 15px;
   border-radius: 5px;
-  background-color: white;
+  background-color: rgb(23, 22, 22);
+  box-sizing: border-box;
+  
+
 }
 
 .nav-link:hover {
-  color: #ffeb3b; 
+  color: green; 
   background-color: rgba(0, 0, 0, 0.2); 
+  font: bold;
 }
