--- client/src/components/TimeEstimates.css
@@ -1,134 +1,0 @@
-/* Main container for the time estimates component */
-.time-estimates {
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-  padding: 2rem;
-  background-color:#bbb4b4; 
-}
-
-/* Box around the time estimates content */
-.time-estimates-box {
-  display: flex;
-  flex-direction: column;
-  justify-content: flex-start;
-  align-items: center;
-  width: 100%;
-  max-width: 900px;
-  padding: 2rem;
-  background-color: white;
-  text-align: left;
-  border-radius: 8px;
-  box-shadow: 0 4px 10px rgba(0, 0, 0, 0.1);
-}
-
-/* Title styling */
-.time-estimates-box h3 {
-  margin-bottom: 1rem;
-  font-size: 1.8rem;
-  color: #333;
-}
-
-/* Route information summary */
-.summary-box {
-  margin-bottom: 2rem;
-  font-size: 1.2rem;
-  color: #333;
-}
-
-/* Styled details (Total Travel Time, Distance) */
-.summary-box div {
-  margin: 0.5rem 0;
-  font-weight: 500;
-}
-
-/* List of ordered locations */
-.locations-list {
-  list-style-type: none;
-  padding: 0;
-  width: 100%;
-  text-align: center;
-}
-
-.locations-list li {
-  background-color: #f9f9f9;
-  margin: 1rem 0;
-  padding: 1rem;
-  border-radius: 5px;
-  box-shadow: 0 2px 4px rgba(0, 0, 0, 0.1);
-  border: 2px solid #d6c3c3;
-  text-align: center;
-}
-
-.locations-list li {
-  margin-bottom: 0.5rem;
-}
-
-.locations-list li strong {
-  color: #007bff;
-}
-
-.locations-list li span {
-  color: #555;
-}
-
-/* Loading and error message */
-.loading-message {
-  font-size: 1.5rem;
-  color: #333;
-  margin: 1rem 0;
-}
-
-/* Map container box style */
-.map-container-box {
-  width: 100%;
-  height: 400px;
-  margin-top: 1rem;
-  border: 2px solid #ccc;
-  border-radius: 10px;
-  background-color: white;
-  box-shadow: 0 4px 10px rgba(0, 0, 0, 0.1);
-}
-
-/* Responsive styles for smaller screens (tablets) */
-@media (max-width: 768px) {
-  .time-estimates-box {
-    width: 95%; /* Make the box take more space on smaller screens */
-    padding: 1rem;
-  }
-
-  .summary-box {
-    font-size: 1rem;
-  }
-
-  .locations-list li {
-    font-size: 1rem;
-    padding: 1rem;
-  }
-
-  .map-container-box {
-    height: 350px; /* Reduce map height on smaller screens */
-  }
-}
-
-/* Responsive styles for very small screens (mobiles) */
-@media (max-width: 480px) {
-  .time-estimates-box {
-    width: 100%;
-    padding: 1rem;
-  }
-
-  .summary-box {
-    font-size: 0.9rem;
-  }
-
-  .locations-list li {
-    font-size: 0.9rem;
-    padding: 0.8rem;
-  }
-
-  .map-container-box {
-    height: 250px; /* Further reduce map height */
-  }
-}
