+++ client/src/components/dashboard.jsx
@@ -1,7 +1,6 @@
+
 import ErrorBoundary from "./ErrorBoundary.jsx";
 import MapDisplay from "./MapDisplay.jsx";
-import TimeEstimates from "./TimeEstimates.jsx";
-import "./Dashboard.css"; 
 
 const RouteDashboard = () => {
   return (
@@ -11,11 +10,9 @@ const RouteDashboard = () => {
         <ErrorBoundary>
           <MapDisplay />
         </ErrorBoundary>
-        <TimeEstimates />
       </div>
     </div>
   );
 };
 
 export default RouteDashboard;
-
